HR Agent App/requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
fastapi
uvicorn
streamlit
